+++ NPacMan.sln
@@ -20,7 +20,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Docs", "Docs", "{70F9337E-F
 		GhostMovement.md = GhostMovement.md
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NPacMan.ConsoleApp", "src\NPacMan.ConsoleApp\NPacMan.ConsoleApp.csproj", "{7E5D00CC-D5A1-4D76-AA27-FFC529691212}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NPacMan.ConsoleApp", "src\NPacMan.ConsoleApp\NPacMan.ConsoleApp.csproj", "{7E5D00CC-D5A1-4D76-AA27-FFC529691212}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
