+++ src/NPacMan.Game/Game.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.CodeDom;
 using System.Collections.Generic;
 using System.Linq;
 using System.Collections.Immutable;
@@ -8,125 +7,6 @@ using Automatonymous;
 
 namespace NPacMan.Game
 {
-    internal class GameState
-    {
-        public GameState(IGameSettings settings)
-        {
-            Lives = settings.InitialLives;
-            Status = settings.InitialGameStatus switch
-            {
-                GameStatus.Initial => nameof(GameStateMachine.Initial),
-                GameStatus.Alive => nameof(GameStateMachine.Alive),
-                GameStatus.Dying => nameof(GameStateMachine.Dying),
-                GameStatus.Respawning => nameof(GameStateMachine.Respawning),
-                GameStatus.Dead => nameof(GameStateMachine.Dead),
-                _ => throw new NotImplementedException($"No map for InitialGameStatus '{settings.InitialGameStatus}")
-            };
-            Score = 0;
-            GhostsVisible = true;
-            TimeToChangeState = null;
-        }
-
-        public string Status { get; set; } = null!;
-
-        public DateTime? TimeToChangeState { get; set; }
-
-        public int Lives { get; set; }
-
-        public bool GhostsVisible { get; set; }
-
-        public int Score { get; set; }
-    }
-
-    internal interface IGameActions
-    {
-        void MoveGhostsHome();
-        void ShowGhosts(GameState gameState);
-        void ScatterGhosts();
-        void GhostToChase();
-        Task MovePacMan(DateTime now);
-        Task MoveGhosts(DateTime now);
-        void HideGhosts(GameState gameState);
-        void MovePacManHome();
-    }
-
-    internal class GameStateMachine :
-               AutomatonymousStateMachine<GameState>
-    {
-        public GameStateMachine(IGameActions game, IGameSettings settings)
-        {
-            InstanceState(x => x.Status);
-
-            Initially(
-                When(Tick)
-                    .Then(context => game.MoveGhostsHome())
-                    .Then(context => game.ShowGhosts(context.Instance))
-                    .Then(context => game.ScatterGhosts())
-                    .Then(context => context.Instance.TimeToChangeState = context.Data.Now.AddSeconds(settings.InitialScatterTimeInSeconds))
-                    .TransitionTo(Scatter));
-
-            During(Scatter,
-                 When(Tick, context => context.Data.Now >= context.Instance.TimeToChangeState)
-                    .Then(context => context.Instance.TimeToChangeState = context.Data.Now.AddSeconds(settings.ChaseTimeInSeconds))
-                     .Then(context => game.GhostToChase())
-                     .TransitionTo(Alive));
-
-            During(Alive,
-                 When(Tick, context => context.Data.Now >= context.Instance.TimeToChangeState)
-                    .Then(context => context.Instance.TimeToChangeState = context.Data.Now.AddSeconds(settings.InitialScatterTimeInSeconds))
-                     .Then(context => game.ScatterGhosts())
-                     .TransitionTo(Scatter));
-
-            During(Scatter, Alive,
-                When(Tick)
-                    .ThenAsync(async context => await game.MoveGhosts(context.Data.Now))
-                    .Then(context => game.MovePacMan(context.Data.Now)),
-                When(CoinEaten)
-                    .Then(context => context.Instance.Score += 10),
-                    //.Then(context => game._soundSet.Chomp()),
-                When(PacManCaughtByGhost)
-                    .Then(context => context.Instance.Lives -= 1)
-                    .Then(context => context.Instance.TimeToChangeState = context.Data.Now.AddSeconds(4))
-                    .TransitionTo(Dying));
-
-            During(Dying,
-                When(Tick, context => context.Data.Now >= context.Instance.TimeToChangeState)
-                    .Then(context => game.HideGhosts(context.Instance))
-                    .Then(context => context.Instance.TimeToChangeState = context.Data.Now.AddSeconds(4))
-                    //.Then(context => game._soundSet.Death())
-                    .IfElse(context => context.Instance.Lives > 0,
-                        binder => binder.TransitionTo(Respawning),
-                        binder => binder.Finalize()));
-
-            During(Respawning,
-                When(Tick, context => context.Data.Now >= context.Instance.TimeToChangeState)
-                    .Then(context => context.Instance.TimeToChangeState = context.Data.Now.AddSeconds(4))
-                    .Then(context => game.MoveGhostsHome())
-                    .Then(context => game.MovePacManHome())
-                    .Then(context => game.ShowGhosts(context.Instance))
-                    .TransitionTo(Alive));
-
-            During(Dead, Ignore(Tick));
-        }
-
-
-        public State Alive { get; private set; } = null!;
-
-        public State Scatter { get; private set; } = null!;
-        public State Dying { get; private set; } = null!;
-        public State Respawning { get; private set; } = null!;
-        public State Dead { get; private set; } = null!;
-        public Event<Tick> Tick { get; private set; } = null!;
-        public Event<Tick> PacManCaughtByGhost { get; private set; } = null!;
-        public Event CoinEaten { get; private set; } = null!;
-    }
-
-
-    internal class Tick
-    {
-        public DateTime Now { get; set; }
-    }
-
     public class Game : IGameActions
     {
         public int Score => _gameState.Score;
