+++ src/NPacMan.Game/GameState.cs
@@ -1,0 +1,34 @@
+﻿using System;
+
+namespace NPacMan.Game
+{
+    internal class GameState
+    {
+        public GameState(IGameSettings settings)
+        {
+            Lives = settings.InitialLives;
+            Status = settings.InitialGameStatus switch
+            {
+                GameStatus.Initial => nameof(GameStateMachine.Initial),
+                GameStatus.Alive => nameof(GameStateMachine.Alive),
+                GameStatus.Dying => nameof(GameStateMachine.Dying),
+                GameStatus.Respawning => nameof(GameStateMachine.Respawning),
+                GameStatus.Dead => nameof(GameStateMachine.Dead),
+                _ => throw new NotImplementedException($"No map for InitialGameStatus '{settings.InitialGameStatus}")
+            };
+            Score = 0;
+            GhostsVisible = true;
+            TimeToChangeState = null;
+        }
+
+        public string Status { get; set; } = null!;
+
+        public DateTime? TimeToChangeState { get; set; }
+
+        public int Lives { get; set; }
+
+        public bool GhostsVisible { get; set; }
+
+        public int Score { get; set; }
+    }
+}
