+++ src/NPacMan.Game/GameStatus.cs
