+++ src/NPacMan.Game/Tick.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace NPacMan.Game
+{
+    internal class Tick
+    {
+        public DateTime Now { get; set; }
+    }
+}
